requirements.txt
streamlit
langchain
langchain-core
langchain-groq
langchain-ollama
pymongo
python-dotenv
rich